refactor: move logic into containers
