+++ src/containers/CommunityMembers.vue
@@ -1,0 +1,64 @@
+<template>
+  <j-flex gap="500" direction="column">
+    <j-text variant="heading"
+      >All group members ({{ community.members.length ?? 0 }})</j-text
+    >
+    <j-input
+      size="lg"
+      placeholder="Search for member"
+      type="search"
+      :value="searchValue"
+      @input="(e) => (searchValue = e.target.value)"
+    >
+      <j-icon name="search" size="sm" slot="start"></j-icon>
+    </j-input>
+    <j-flex wrap gap="600">
+      <j-flex
+        gap="300"
+        v-for="communityMember in filteredCommunityMemberList"
+        :key="communityMember.data.profile['foaf:AccountName']"
+        inline
+        direction="column"
+        a="center"
+      >
+        <j-avatar
+          size="lg"
+          :src="
+            communityMember.data.profile['schema:image']
+              ? JSON.parse(communityMember.data.profile['schema:image'])[
+                  'schema:contentUrl'
+                ]
+              : require('@/assets/images/avatar-placeholder.png')
+          "
+        />
+        <j-text variant="body">{{
+          communityMember.data.profile["foaf:AccountName"]
+        }}</j-text>
+      </j-flex>
+    </j-flex>
+  </j-flex>
+</template>
+
+<script lang="ts">
+import { defineComponent } from "vue";
+import Expression from "@perspect3vism/ad4m/Expression";
+
+export default defineComponent({
+  data() {
+    return {
+      searchValue: "",
+    };
+  },
+  computed: {
+    community() {
+      return this.$store.getters.getCommunity(this.$route.params.communityId);
+    },
+    filteredCommunityMemberList(): Expression[] {
+      const members: Expression[] = this.community.members;
+      return members.filter((m: Expression) =>
+        Object(m.data).profile["foaf:AccountName"].includes(this.searchValue)
+      );
+    },
+  },
+});
+</script>
