+++ src/views/community-view/community-sidebar/CommunitySidebar.vue
@@ -1,5 +1,5 @@
 <template>
-  <div class="community-sidebar" v-if="community != null">
+  <div class="community-sidebar" v-if="community">
     <button class="community-sidebar__header">
       <j-flex j="between" gap="300">
         <j-flex a="center" gap="400">
@@ -8,10 +8,22 @@
             :src="require('@/assets/images/junto_app_icon.png')"
           />
           <div>
-            <j-text weight="500" color="ui-800" nomargin size="500">
+            <j-text
+              v-if="community.name"
+              weight="500"
+              color="ui-800"
+              nomargin
+              size="500"
+            >
               {{ community.name }}
             </j-text>
-            <j-text weight="300" color="ui-800" nomargin size="400">
+            <j-text
+              v-if="community.description"
+              weight="300"
+              color="ui-800"
+              nomargin
+              size="400"
+            >
               {{ community.description }}
             </j-text>
           </div>
@@ -89,100 +101,27 @@
       :open="showGroupMembers"
       @toggle="(e) => (showGroupMembers = e.target.open)"
     >
-      <j-flex gap="500" direction="column">
-        <j-text variant="heading"
-          >All group members ({{ community.members.length ?? 0 }})</j-text
-        >
-        <j-input
-          placeholder="Search for member"
-          type="search"
-          :value="searchValue"
-          @input="(e) => (searchValue = e.target.value)"
-        ></j-input>
-        <j-flex wrap gap="600">
-          <j-flex
-            gap="300"
-            v-for="communityMember in filteredCommunityMemberList"
-            :key="communityMember.data.profile['foaf:AccountName']"
-            inline
-            direction="column"
-            a="center"
-          >
-            <j-avatar
-              size="lg"
-              :src="
-                communityMember.data.profile['schema:image']
-                  ? JSON.parse(communityMember.data.profile['schema:image'])[
-                      'schema:contentUrl'
-                    ]
-                  : require('@/assets/images/avatar-placeholder.png')
-              "
-            />
-            <j-text variant="body">{{
-              communityMember.data.profile["foaf:AccountName"]
-            }}</j-text>
-          </j-flex>
-        </j-flex>
-      </j-flex>
+      <community-members />
     </j-modal>
 
     <j-modal
       :open="showUpdateCommunity"
       @toggle="(e) => (showUpdateCommunity = e.target.open)"
     >
-      <j-text variant="heading">Edit Community</j-text>
-      <j-flex direction="column" gap="400">
-        <j-input
-          size="lg"
-          label="Name"
-          :value="communityName"
-          @input="(e) => (communityName = e.target.value)"
-        ></j-input>
-        <j-input
-          size="lg"
-          label="Description"
-          :value="communityDescription"
-          @keydown.enter="updateCommunity"
-          @input="(e) => (communityDescription = e.target.value)"
-        ></j-input>
-        <j-button
-          size="lg"
-          :loading="isUpdatingCommunity"
-          :disabled="isUpdatingCommunity"
-          @click="updateCommunity"
-          variant="primary"
-        >
-          Save
-        </j-button>
-      </j-flex>
+      <edit-community
+        @submit="showUpdateCommunity = false"
+        @cancel="showUpdateCommunity = false"
+      />
     </j-modal>
 
     <j-modal
       :open="showCreateChannel"
       @toggle="(e) => (showCreateChannel = e.target.open)"
     >
-      <j-text variant="heading">Create Channel</j-text>
-      <j-text variant="body">
-        Channels are ways to organize your conversations by topics.
-      </j-text>
-      <j-flex direction="column" gap="400">
-        <j-input
-          size="lg"
-          label="Name"
-          :value="channelName"
-          @keydown.enter="createChannel"
-          @input="(e) => (channelName = e.target.value)"
-        ></j-input>
-        <j-button
-          size="lg"
-          :loading="isCreatingChannel"
-          :disabled="isCreatingChannel"
-          @click="createChannel"
-          variant="primary"
-        >
-          Create Channel
-        </j-button>
-      </j-flex>
+      <create-channel
+        @submit="showCreateChannel = false"
+        @cancel="showCreateChannel = false"
+      />
     </j-modal>
   </div>
 </template>
@@ -190,34 +129,18 @@
 <script lang="ts">
 import { defineComponent } from "vue";
 import AvatarGroup from "@/components/avatar-group/AvatarGroup.vue";
-import Expression from "@perspect3vism/ad4m/Expression";
+import EditCommunity from "@/containers/EditCommunity.vue";
+import CreateChannel from "@/containers/CreateChannel.vue";
+import CommunityMembers from "@/containers/CommunityMembers.vue";
 
 export default defineComponent({
-  components: { AvatarGroup },
+  components: { AvatarGroup, EditCommunity, CreateChannel, CommunityMembers },
   props: ["community"],
-  watch: {
-    community: {
-      handler: function ({ name, description }) {
-        this.communityName = name;
-        this.communityDescription = description;
-
-        this.searchValue = "";
-      },
-      deep: true,
-      immediate: true,
-    },
-  },
   data() {
     return {
-      isCreatingChannel: false,
-      channelName: "",
       showCreateChannel: false,
-      isUpdatingCommunity: false,
-      communityName: "",
-      communityDescription: "",
       showUpdateCommunity: false,
       showGroupMembers: false,
-      searchValue: "",
     };
   },
   methods: {
@@ -234,47 +157,6 @@ export default defineComponent({
       this.$store.commit("showSuccessToast", {
         message: "Your custom invite code is copied to your clipboard!",
       });
-    },
-    async updateCommunity() {
-      const { communityId } = this.$route.params;
-      this.isUpdatingCommunity = true;
-      this.$store
-        .dispatch("updateCommunity", {
-          communityId: communityId,
-          name: this.communityName,
-          description: this.communityDescription,
-        })
-        .then(() => {
-          this.showUpdateCommunity = false;
-        })
-        .finally(() => {
-          this.isUpdatingCommunity = false;
-        });
-    },
-    async createChannel() {
-      const { communityId } = this.$route.params;
-      const name = this.channelName;
-      this.isCreatingChannel = true;
-      this.$store
-        .dispatch("createChannel", {
-          communityId,
-          name,
-        })
-        .then(() => {
-          this.showCreateChannel = false;
-          this.channelName = "";
-        })
-        .finally(() => {
-          this.isCreatingChannel = false;
-        });
-    },
-  },
-  computed: {
-    filteredCommunityMemberList(): Expression[] {
-      const members: Expression[] = this.community.members;
-      return members.filter((m: Expression) =>
-        Object(m.data).profile["foaf:AccountName"].includes(this.searchValue)
-      );
     },
   },
 });
