+++ src/views/main-view/main-sidebar/BottomSection.vue
@@ -32,13 +32,26 @@
           <j-text nomargin>{{ userProfile.username }}</j-text>
         </j-flex>
         <j-menu-item @click="isEditProfileOpen = true">
+          <j-icon size="sm" slot="start" name="pencil"></j-icon>
           Edit profile
         </j-menu-item>
-        <j-menu-item>View profile</j-menu-item>
-        <j-menu-item @click="updateApp.func">{{ updateApp.text }}</j-menu-item>
-        <j-menu-item @click="isSettingsOpen = true">Settings</j-menu-item>
+        <j-menu-item>
+          <j-icon size="sm" slot="start" name="person"></j-icon>
+          View profile
+        </j-menu-item>
+        <j-menu-item @click="updateApp.func">
+          <j-icon size="sm" slot="start" name="cloud-download"></j-icon>
+          {{ updateApp.text }}
+        </j-menu-item>
+        <j-menu-item @click="isSettingsOpen = true">
+          <j-icon size="sm" slot="start" name="gear"></j-icon>
+          Settings
+        </j-menu-item>
         <router-link :to="{ name: 'signup' }" v-slot="{ navigate }">
-          <j-menu-item @click="navigate">Log out</j-menu-item>
+          <j-menu-item @click="navigate">
+            <j-icon size="sm" slot="start" name="door-closed"></j-icon>
+            Log out
+          </j-menu-item>
         </router-link>
       </j-menu>
     </j-popover>
@@ -47,23 +60,10 @@
       :open="isEditProfileOpen"
       @toggle="(e) => (isEditProfileOpen = e.target.open)"
     >
-      <j-flex direction="column" gap="700">
-        <j-text variant="heading">Edit profile</j-text>
-        <avatar-upload
-          :value="profilePicture"
-          @change="(url) => (profilePicture = url)"
-        ></avatar-upload>
-        <j-input
-          size="lg"
-          label="Username"
-          @keydown.enter="updateUser"
-          :value="userProfile?.username"
-          @input="(e) => (username = e.target.value)"
-        ></j-input>
-        <j-button size="lg" variant="primary" @click="updateUser" full>
-          Save
-        </j-button>
-      </j-flex>
+      <edit-profile
+        @submit="isEditProfileOpen = false"
+        @cancel="isEditProfileOpen = false"
+      />
     </j-modal>
 
     <j-modal
@@ -71,67 +71,22 @@
       :open="isSettingsOpen"
       @toggle="(e) => (isSettingsOpen = e.target.open)"
     >
-      <j-flex direction="column" gap="700">
-        <j-text variant="heading">Settings</j-text>
-        <j-flex a="center" j="between">
-          <j-text variant="label">Font family</j-text>
-          <j-tabs
-            :value="fontFamily"
-            @change="(e) => (fontFamily = e.target.value)"
-          >
-            <j-tab-item value="default">Default</j-tab-item>
-            <j-tab-item value="system">System</j-tab-item>
-            <j-tab-item value="monospace">Monospace</j-tab-item>
-          </j-tabs>
-        </j-flex>
-        <j-flex a="center" j="between">
-          <j-text variant="label">Mode</j-text>
-          <j-tabs
-            :value="themeName"
-            @change="(e) => (themeName = e.target.value)"
-          >
-            <j-tab-item value="light">Light</j-tab-item>
-            <j-tab-item value="dark">Dark</j-tab-item>
-          </j-tabs>
-        </j-flex>
-        <j-flex a="center" j="between">
-          <j-text variant="label">Primary color</j-text>
-          <div class="colors">
-            <button
-              v-for="n in [0, 50, 100, 150, 200, 250, 270, 300]"
-              :key="n"
-              class="color-button"
-              :class="{ 'color-button--active': themeHue === n }"
-              @click="() => (themeHue = n)"
-              :style="`--hue: ${n}`"
-            ></button>
-          </div>
-        </j-flex>
-        <j-flex a="center" j="between">
-          <j-text variant="label">Clear State</j-text>
-          <j-button size="md" variant="primary" @click="cleanState">
-            <j-icon size="sm" name="trash"></j-icon>
-            Clear state
-          </j-button>
-        </j-flex>
-        <div>
-          <j-button size="lg" @click="isSettingsOpen = false">Cancel</j-button>
-          <j-button size="lg" variant="primary" @click="updateTheme">
-            Save
-          </j-button>
-        </div>
-      </j-flex>
+      <settings
+        @submit="isSettingsOpen = false"
+        @cancel="isSettingsOpen = false"
+      />
     </j-modal>
   </div>
 </template>
 
 <script lang="ts">
 import { defineComponent, onBeforeMount } from "vue";
-import { Profile, ThemeState } from "@/store";
-import AvatarUpload from "@/components/avatar-upload/AvatarUpload.vue";
+import Settings from "@/containers/Settings.vue";
+import EditProfile from "@/containers/EditProfile.vue";
+import { Profile } from "@/store";
 
 export default defineComponent({
-  components: { AvatarUpload },
+  components: { Settings, EditProfile },
   setup() {
     onBeforeMount(() => {
       window.api.receive("getCleanState", (data: string) => {
@@ -143,100 +98,12 @@ export default defineComponent({
   },
   data() {
     return {
-      hue: getComputedStyle(document.documentElement).getPropertyValue(
-        "--j-color-primary-hue"
-      ),
-      fontFamily: getComputedStyle(document.documentElement).getPropertyValue(
-        "--j-font-family"
-      ),
       isSettingsOpen: false,
       isEditProfileOpen: false,
-      isUpdatingProfile: false,
-      username: "",
-      profilePicture: "",
-      themeName: "",
-      themeHue: "",
     };
   },
-  watch: {
-    isSettingsOpen: function (val) {
-      if (!val) {
-        this.resetTheme();
-      }
-    },
-    themeHue: function (val) {
-      document.documentElement.style.setProperty("--j-color-primary-hue", val);
-    },
-    themeName: function (val) {
-      document.documentElement.setAttribute("theme", val);
-    },
-    fontFamily: function (val: "system" | "default") {
-      const font = {
-        default: `"Avenir", sans-serif`,
-        monospace: `monospace`,
-        system: `-apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Helvetica, Arial, sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol"`,
-      };
-      document.documentElement.style.setProperty("--j-font-family", font[val]);
-    },
-    "userProfile.profilePicture": {
-      handler: function (val: string): void {
-        this.profilePicture = val;
-      },
-      immediate: true,
-    },
-    "userProfile.username": {
-      handler: function (val: string): void {
-        this.username = val;
-      },
-      immediate: true,
-    },
-    "theme.name": {
-      handler: function (val: string): void {
-        this.themeName = val;
-      },
-      immediate: true,
-    },
-    "theme.hue": {
-      handler: function (val: string): void {
-        this.themeHue = val;
-      },
-      immediate: true,
-    },
-    "theme.fontFamily": {
-      handler: function (val: string): void {
-        this.fontFamily = val;
-      },
-      immediate: true,
-    },
-  },
+
   methods: {
-    resetTheme() {
-      this.$store.commit("setTheme", {
-        name: this.theme.name,
-        hue: this.theme.hue,
-      });
-    },
-    updateTheme() {
-      this.$store.commit("setTheme", {
-        name: this.themeName,
-        hue: this.themeHue,
-      });
-      this.isSettingsOpen = false;
-    },
-    updateUser() {
-      this.isUpdatingProfile = true;
-      this.$store
-        .dispatch("updateUser", {
-          username: this.username,
-          profilePicture: this.profilePicture,
-        })
-        .then(() => {
-          this.isEditProfileOpen = false;
-        })
-        .finally(() => {
-          this.isUpdatingProfile = false;
-        });
-    },
     cleanState() {
       window.api.send("cleanState");
     },
@@ -253,9 +120,6 @@ export default defineComponent({
     },
   },
   computed: {
-    theme(): ThemeState {
-      return this.$store.state.ui.theme;
-    },
     userProfile(): Profile {
       return this.$store.state.userProfile || {};
     },
@@ -299,21 +163,5 @@ export default defineComponent({
   gap: var(--j-space-400);
   flex-direction: column;
   align-items: center;
-}
-.color-button {
-  --hue: 0;
-  width: var(--j-element-md);
-  height: var(--j-element-md);
-  background-color: hsl(var(--hue), 100%, 50%);
-  border: 2px solid transparent;
-  outline: 0;
-  border-radius: var(--j-border-radius);
-  margin-right: var(--j-space-200);
-}
-.color-button--active {
-  border-color: var(--j-color-primary-600);
-}
-.colors {
-  max-width: 400px;
 }
 </style>
