+++ src/views/main-view/main-sidebar/CommunitiesList.vue
@@ -1,7 +1,7 @@
 <template>
   <div class="left-nav__communities-list">
     <j-tooltip
-      v-for="community in getCommunities"
+      v-for="community in communities"
       :key="community.perspective"
       :title="community.name"
     >
@@ -14,7 +14,7 @@
       ></j-avatar>
     </j-tooltip>
 
-    <j-tooltip title="Create comminuty">
+    <j-tooltip title="Create community">
       <j-button
         @click="showModal = true"
         variant="primary"
@@ -27,119 +27,26 @@
     </j-tooltip>
 
     <j-modal :open="showModal" @toggle="(e) => (showModal = e.target.open)">
-      <j-flex direction="column" gap="700">
-        <div v-if="tabView === 'Create'">
-          <j-text variant="heading">Create a community </j-text>
-        </div>
-        <div v-if="tabView === 'Join'">
-          <j-text variant="heading">Join a community </j-text>
-          <j-text color="danger-500">
-            This is an early version of Junto with a remote code execution
-            vulnerability. DO NOT join any community from somebody you do not
-            know as its possible for people to use fake joining links to hack
-            you!
-          </j-text>
-        </div>
-
-        <j-tabs
-          size="lg"
-          :value="tabView"
-          @change="(e) => (tabView = e.target.value)"
-        >
-          <j-tab-item>Create</j-tab-item>
-          <j-tab-item>Join</j-tab-item>
-        </j-tabs>
-        <j-flex direction="column" gap="500" v-if="tabView === 'Create'">
-          <j-input
-            size="lg"
-            label="Name"
-            @input="(e) => (newCommunityName = e.target.value)"
-            :value="newCommunityName"
-          ></j-input>
-          <j-input
-            size="lg"
-            label="Description"
-            :value="newCommunityDesc"
-            @keydown.enter="createCommunity"
-            @input="(e) => (newCommunityDesc = e.target.value)"
-          ></j-input>
-          <j-button
-            :loading="isCreatingCommunity"
-            :disabled="isCreatingCommunity"
-            size="lg"
-            full
-            variant="primary"
-            @click="createCommunity"
-          >
-            Create Community
-          </j-button>
-        </j-flex>
-        <j-flex direction="column" gap="200" v-if="tabView === 'Join'">
-          <j-input
-            :value="joiningLink"
-            @input="(e) => (joiningLink = e.target.value)"
-            size="lg"
-            label="Invite link"
-          ></j-input>
-          <j-button
-            :disabled="isJoiningCommunity"
-            :loading="isJoiningCommunity"
-            @click="joinCommunity"
-            size="lg"
-            full
-            variant="primary"
-          >
-            Join Community
-          </j-button>
-        </j-flex>
-      </j-flex>
+      <create-community
+        @submit="showModal = false"
+        @cancel="showModal = false"
+      />
     </j-modal>
   </div>
 </template>
 
 <script lang="ts">
 import { defineComponent } from "vue";
+import CreateCommunity from "@/containers/CreateCommunity.vue";
 
 export default defineComponent({
+  components: { CreateCommunity },
   data() {
     return {
-      tabView: "Create",
-      joiningLink: "",
-      newCommunityName: "",
-      newCommunityDesc: "",
       showModal: false,
-      isJoiningCommunity: false,
-      isCreatingCommunity: false,
     };
   },
   methods: {
-    joinCommunity() {
-      this.isJoiningCommunity = true;
-      this.$store
-        .dispatch("joinCommunity", { joiningLink: this.joiningLink })
-        .then(() => {
-          this.showModal = false;
-        })
-        .finally(() => {
-          this.isJoiningCommunity = false;
-        });
-    },
-    createCommunity() {
-      this.isCreatingCommunity = true;
-      this.$store
-        .dispatch("createCommunity", {
-          perspectiveName: this.newCommunityName,
-          description: this.newCommunityDesc,
-        })
-        .then(() => {
-          this.showModal = false;
-          this.newCommunityName = "";
-          this.newCommunityDesc = "";
-        })
-        .finally(() => {
-          this.isCreatingCommunity = false;
-        });
-    },
     handleCommunityClick(communityId: string) {
       this.$router.push({ name: "community", params: { communityId } });
       if (this.communityIsActive(communityId)) {
@@ -150,9 +57,8 @@ export default defineComponent({
     },
   },
   computed: {
-    getCommunities() {
-      const communities = this.$store.getters.getCommunities;
-      return communities;
+    communities() {
+      return this.$store.state.communities;
     },
     communityIsActive() {
       return (id: string) => this.$route.params.communityId === id;
